refactoring to clean up code after host-remote merge
