+++ app/modules/users/user.ts
@@ -6,14 +6,25 @@ class User {
         this.userId = userId;
         this.username = username;
     }
+
+
 }
 
 class UserManager {
     private _users: Map<string | undefined, User>;
     private _nullUser: User = new User("null", "null");
+    private static _instance: UserManager;
 
-    public constructor() {
+    private constructor() {
         this._users = new Map<string, User>();
+    }
+
+    public static getInstance(): UserManager {
+        if (UserManager._instance == null) {
+            UserManager._instance = new UserManager();
+        }
+
+        return UserManager._instance;
     }
 
     public createUser(userId: string, username: string): User {
